Change padding and squishing. We weren't taking into account the fact that the axis label font descent was being subtracted off when computing axis sizes.
Additionally, I changed the glyphBoxes code for
both horizAxis and verticalAxis so that only the
right-most and top-most tick labels are accounted
for respectively.  There needs to be a big change
here.  We should be using information about the
plot label width/height and the axis padding when
squishing the draw area.  This needs to be done
by the plot itself, not by each axis.