Clarify vg.Canvas.Push() and vg.Canvas.Pop().
They save/restore the line width, dash pattern, transforms,
and color.  There was also a slight bug in vgimg where its
Canvas cached the current color so that it could be used to
color text, but the cached color did not update with Push
and Pop.  This now works properly.  Fixes issue 54.