Add the FontDir variable to vg.
FontDir is a slice of directories searched for the font data files. It
is initialized with the VGFONTPATH environment variable and with the
source code import path to the vg/fonts directory if either or both
are defined. Othewise it contains just the current directory. A user
can change this variable if they want to place fonts in a separate
location.

Fixes issue 89.