Improve error bar documentation and usage.
Added comments documenting how the NewXErrorBars and NewYErrorBars
interpret the error values returned by the Errorer interfaces. Also
changed how the errors are interpreted: the values are relative to the
corresponding X or Y point, the absolute value of the error is taken,
the first is subtracted and the second is added. This is backward
compatible with any previously-correct usage of the error bars plotter
(which assumed that both error vaules were positive) and it removes
any ambiguity about whether the error values should be negative?it
doesn't matter now.

Fixes issue 141.